+++ enterprise/src/main/java/com/foodtraceability/enterprise/entity/TechTemplate.java
@@ -7,7 +7,7 @@ import com.baomidou.mybatisplus.annotation.TableName;
 @TableName("t_tech_template")
 public class TechTemplate {
     @TableId(type = IdType.AUTO)
-    private Integer templateId;
+    private Long templateId;
     private String templateName;
     private String version;
     private String applicableProduct;
@@ -28,8 +28,8 @@ public class TechTemplate {
     private String updateTime;
     private int isDeleted;
 
-    public Integer getTemplateId() { return templateId; }
-    public void setTemplateId(Integer templateId) { this.templateId = templateId; }
+    public Long getTemplateId() { return templateId; }
+    public void setTemplateId(Long templateId) { this.templateId = templateId; }
 
     public String getTemplateName() { return templateName; }
     public void setTemplateName(String templateName) { this.templateName = templateName; }
